+++ packages/g-plugin-zdog-canvas-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @antv/g-plugin-zdog-canvas-renderer
 
+## 2.0.43
+
+### Patch Changes
+
+-   @antv/g-canvas@2.0.44
+
 ## 2.0.42
 
 ### Patch Changes