+++ packages/g-webgl/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-webgl
 
+## 2.0.48
+
+### Patch Changes
+
+-   Updated dependencies [76644e3]
+    -   @antv/g-plugin-image-loader@2.1.23
+    -   @antv/g-plugin-device-renderer@2.2.23
+
 ## 2.0.47
 
 ### Patch Changes